refactor: Formal code cleaning
Implement linter suggestions
Both manual and automatic (fix-linter) alterations addressing linter
warnings.
Reverted fix-linter's renaming of getRegisterAccessor_impl to getRegisterAccessorImpl
back to getRegisterAccessor_impl, since the alteration won't compile,
and corresponding linter warnings are suppressed.
In SerialPort.cc, bit operations now only occur with uints
Change to size_t in SerialPort::send()
Eliminates C-style array in SerialPort::readline()
Satisfies the linter for the move semantics of the DummyServer::LockingString
constructor.

+++ tests/testSerialPort.cc
@@ -22,7 +22,7 @@ struct SerialCommandHandlerFixture {
 BOOST_FIXTURE_TEST_SUITE(SerialCommandHandlerTests, SerialCommandHandlerFixture)
 
 BOOST_AUTO_TEST_CASE(testSimpleReply) {
-  for(long l = 0; l < 10; l++) {
+  for(int32_t l = 0; l < 10; l++) {
     std::string cmd = "reply" + std::to_string(l);
     s.write(cmd);
     std::string res = s.waitAndReadline();
